notification-service/requirements.txt
pika==1.3.2
boto3==1.35.0
SQLAlchemy==2.0.36
Jinja2==3.1.4
python-dotenv==1.0.1
